@shipilev 2021-01-28
> I thought fetch_and_add is pretty cheap now, and claiming one thread at a time can balance workload better ...

Atomics are cheap when uncontended. When per-thread work is small, we run into contended atomic, and get unnecessary slowdowns. Chunking might indeed make balancing less precise, but we might not care as much as contending with many GC worker threads.